Moved `pg-mem` helper functions into `src/testHelpers.ts`
